requirements.txt
aniso8601==2.0.0
asn1crypto==0.24.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorhash==1.0.2
configparser==3.5.0
cryptography==2.1.4
cx-Oracle==7.0.0
dateutils==0.6.6
decorator==4.2.1
et-xmlfile==1.0.1
Flask==0.12.2
flask-ldap3-login==0.9.13
Flask-Mail==0.9.1
flask-monitoring-dashboard==1.9
Flask-RESTful==0.3.6
flask-restplus==0.10.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
get==0.0.39
gevent==1.2.2
greenlet==0.4.13
grequests==0.3.0
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
JayDeBeApi3==1.3.2
jdcal==1.3
Jinja2==2.10
JPype1-py3==0.5.5.2
jsonschema==2.6.0
jupyter-core==4.4.0
ldap3==2.4.1
MarkupSafe==1.0
nbformat==4.4.0
openpyxl==2.5.0
paramiko==2.4.0
plotly==2.5.0
post==0.0.26
psutil==5.4.3
public==0.0.65
py4j==0.10.6
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.7.0
pytz==2017.3
query-string==0.0.28
request==0.0.26
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
suds-jurko==0.6
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
